requirements.txt
virtualenv
asgiref==3.2.3
Django==3.0.1
Pillow==6.2.1
pytz==2019.3
sqlparse==0.3.0


3.superuser
user: sodaiadmin
pass:sodaiadmin
email:sodaiadmin@gmail.com

